- Fixed a bug in Config where the type of strings was not correctly recognized in Python 2. Fixed using the "six" library, released under MIT license. - Renamed ParamsScheme to ParamScheme - Renamed fulfilment_automation.py to fulfillment_automation.py
