+++ requirements.txt
@@ -3,4 +3,5 @@ chardet==3.0.4
 idna==2.8
 marshmallow==2.18.0
 requests==2.21.0
+six==1.12.0
 urllib3==1.24.1
